# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * BIRU MENARA API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AdminUser {
  isAdmin: boolean;
}

export interface AdminCredentials {
  password: string;
}

export interface Announcement {
  id: number;
  title: string;
  content: string;
  date: string;
  urgent: boolean;
  published: boolean;
  createdAt: string;
}

export interface AnnouncementInput {
  title: string;
  content: string;
  date: string;
  urgent?: boolean;
  published?: boolean;
}

export interface AnnouncementUpdate {
  title?: string;
  content?: string;
  date?: string;
  urgent?: boolean;
  published?: boolean;
}

export interface NewsPost {
  id: number;
  title: string;
  excerpt: string;
  content: string;
  date: string;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
  published: boolean;
  createdAt: string;
}

export interface NewsPostInput {
  title: string;
  excerpt: string;
  content: string;
  date: string;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
  published?: boolean;
}

export interface NewsPostUpdate {
  title?: string;
  excerpt?: string;
  content?: string;
  date?: string;
  category?: string;
  /** @nullable */
  imageUrl?: string | null;
  published?: boolean;
}

export interface ScheduleMatch {
  id: number;
  sport: string;
  gender: string;
  level: string;
  venue: string;
  date: string;
  time: string;
  status: string;
  /** @nullable */
  result?: string | null;
  sortOrder: number;
}

export interface ScheduleMatchInput {
  sport: string;
  gender: string;
  level: string;
  venue: string;
  date: string;
  time: string;
  status: string;
  /** @nullable */
  result?: string | null;
  sortOrder?: number;
}

export interface ScheduleMatchUpdate {
  sport?: string;
  gender?: string;
  level?: string;
  venue?: string;
  date?: string;
  time?: string;
  status?: string;
  /** @nullable */
  result?: string | null;
  sortOrder?: number;
}

export interface TeamMember {
  id: number;
  role: string;
  name: string;
  memberType: string;
  sortOrder: number;
  /** @nullable */
  imageUrl?: string | null;
}

export interface TeamMemberInput {
  role: string;
  name: string;
  memberType: string;
  sortOrder?: number;
  /** @nullable */
  imageUrl?: string | null;
}

export interface TeamMemberUpdate {
  role?: string;
  name?: string;
  memberType?: string;
  sortOrder?: number;
  /** @nullable */
  imageUrl?: string | null;
}

export interface DownloadFile {
  id: number;
  name: string;
  category: string;
  size: string;
  fileUrl: string;
  sortOrder: number;
}

export interface DownloadFileInput {
  name: string;
  category: string;
  size: string;
  fileUrl: string;
  sortOrder?: number;
}

export interface DownloadFileUpdate {
  name?: string;
  category?: string;
  size?: string;
  fileUrl?: string;
  sortOrder?: number;
}

export interface SiteSettings {
  id: number;
  instagram: string;
  facebook: string;
  address: string;
  /** @nullable */
  phone?: string | null;
  heroTitle: string;
  heroSlogan: string;
  heroSubSlogan: string;
}

export interface SiteSettingsUpdate {
  instagram?: string;
  facebook?: string;
  address?: string;
  /** @nullable */
  phone?: string | null;
  heroTitle?: string;
  heroSlogan?: string;
  heroSubSlogan?: string;
}