# lib/api-zod/src/generated/types/announcement.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * BIRU MENARA API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Announcement {
  id: number;
  title: string;
  content: string;
  date: string;
  urgent: boolean;
  published: boolean;
  createdAt: string;
}