# lib/api-zod/src/generated/types/announcementUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * BIRU MENARA API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AnnouncementUpdate {
  title?: string;
  content?: string;
  date?: string;
  urgent?: boolean;
  published?: boolean;
}