# lib/api-zod/src/generated/types/downloadFileInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * BIRU MENARA API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DownloadFileInput {
  name: string;
  category: string;
  size: string;
  fileUrl: string;
  sortOrder?: number;
}