# lib/api-zod/src/generated/types/newsPost.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * BIRU MENARA API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NewsPost {
  id: number;
  title: string;
  excerpt: string;
  content: string;
  date: string;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
  published: boolean;
  createdAt: string;
}